acomodamos el modulo a su lugar, y agregamos la ruta a las importaciones
