+++ src/components/ButtonUser.astro
@@ -1,56 +1,48 @@
 ---
-import { LandingPages, UserPages } from "@constants/pages";
-import { AccountCircle } from "@landingPage/userinformation/iconos/Account";
+import { LandingPages, UserPages } from '@constants/pages'
+import { AccountCircle } from 'src/modules/userinformation/components/iconos/Account'
 
 interface Props {
-  class?: string;
-  user: string;
-  email: string;
+	class?: string
+	user: string
+	email: string
 }
-const { user, email, class: className } = Astro.props;
+const { user, email, class: className } = Astro.props
 ---
 
-<div class="grid">
-  <div class="">
-    <button
-      class={"relative px-5 py-3 text-3xl rounded-lg group text-black dark:text-white dark:bg-gray-50" +
-        className}
-      ><AccountCircle />
+<div class='grid'>
+	<div class=''>
+		<button
+			class={'relative px-5 py-3 text-3xl rounded-lg group text-black dark:text-white dark:bg-gray-50' + className}
+			><AccountCircle />
 
-      <div
-        class="absolute w-80 mt-1 top-full right-0 rounded-lg text-left bg-white text-gray-700 dark:bg-black dark:text-white pt-1 shadow-md scale-y-0 group-focus:scale-y-100 origin-top duration-200"
-      >
-        <div
-          class="flex flex-col p-5 space-y-3 text-lg font-extrabold opacity-70 dark:text-white dark:opacity-80"
-        >
-          <h3 class="">
-            {user}
-          </h3>
-          <h3 class="">
-            {email}
-          </h3>
-        </div>
-        <div class="flex flex-col p-2 text-center text-xl">
-          {
-            Object.values(UserPages).map((page) => (
-              <a
-                href={page.url}
-                class={
-                  "relative p-2 py-4 z-10 block hover:bg-gray-100 dark:hover:bg-gray-500"
-                }
-              >
-                {page.title}
-              </a>
-            ))
-          }
-          <a href={LandingPages.Home.url}>
-            <span
-              class="relative p-2 py-4 z-10 block hover:bg-gray-100 dark:hover:bg-gray-500"
-              >Log Out</span
-            >
-          </a>
-        </div>
-      </div>
-    </button>
-  </div>
+			<div
+				class='absolute w-80 mt-1 top-full right-0 rounded-lg text-left bg-white text-gray-700 dark:bg-black dark:text-white pt-1 shadow-md scale-y-0 group-focus:scale-y-100 origin-top duration-200'
+			>
+				<div class='flex flex-col p-5 space-y-3 text-lg font-extrabold opacity-70 dark:text-white dark:opacity-80'>
+					<h3 class=''>
+						{user}
+					</h3>
+					<h3 class=''>
+						{email}
+					</h3>
+				</div>
+				<div class='flex flex-col p-2 text-center text-xl'>
+					{
+						Object.values(UserPages).map((page) => (
+							<a
+								href={page.url}
+								class={'relative p-2 py-4 z-10 block hover:bg-gray-100 dark:hover:bg-gray-500'}
+							>
+								{page.title}
+							</a>
+						))
+					}
+					<a href={LandingPages.Home.url}>
+						<span class='relative p-2 py-4 z-10 block hover:bg-gray-100 dark:hover:bg-gray-500'>Log Out</span>
+					</a>
+				</div>
+			</div>
+		</button>
+	</div>
 </div>
