--- src/modules/landingPage/userinformation/BasicProfile.astro
@@ -1,38 +1,0 @@
----
-import { AccountCircle } from "./iconos/Account";
-import Edit_Input from "@landingPage/userinformation/components/Edit_Input";
-import React from "react";
-
-interface Props {
-  class?: string;
-  user: string;
-  name: string;
-  lastname: string;
-  email: string;
-}
-const { user, name, lastname, email, class: className } = Astro.props;
----
-
-<div
-  class="grid grid-cols-1 rounded-xl bg-component border border-black/10 dark:bg-component-dark dark:border-white/30 shadow-md p-5 justify-start gap-4"
->
-  <div>
-    <AccountCircle />
-    <div class="flex flex-col text-xl dark:text-white space-y-2">
-      <a class="flex font-semibold gap-10 items-center"
-        >Nombre de Usuario:<h3 class="ml-2 opacity-80">{name} {lastname}</h3>
-      </a>
-
-      <a class="flex font-semibold gap-40 items-center">
-        Usuario:
-        <h3 class="ml-1">
-          <Edit_Input id="1" text={user} .string client:load />
-        </h3>
-      </a>
-      <a class="flex font-semibold gap-14 items-center">
-        Correo Electrónico:
-        <h3><Edit_Input id="2" text={email} client:load /></h3>
-      </a>
-    </div>
-  </div>
-</div>
