+++ src/pages/account/profile.astro
@@ -1,36 +1,31 @@
 ---
-import { LandingPages } from "@constants/pages";
-import LandingPage from "@landingPage/layout/LandingPage.astro";
-import BasicProfile from "@landingPage/userinformation/BasicProfile.astro";
-import PersonalInformation from "@landingPage/userinformation/PersonalInformation.astro";
-import MateriasInformation from "@landingPage/userinformation/MateriasInformation.astro";
-
-import React, { createContext, useState } from "react";
-
-// export const EditContext = createContext();
-
-// const [editingComponentId, setEditingComponentId] = useState(null);
+import { LandingPages } from '@constants/pages'
+import LandingPage from '@landingPage/layout/LandingPage.astro'
+import BasicProfile from '@userinformation/components/BasicProfile.astro'
+import MateriasInformation from '@userinformation/components/MateriasInformation.astro'
+import PersonalInformation from '@userinformation/components/PersonalInformation.astro'
 ---
 
-<LandingPage title="EduVision" page={LandingPages.Home}>
-  <h1 class="text-5xl font-extrabold dark:text-white">Mi Perfil</h1>
-  <!-- <EditContext.Provider value={{ editingComponentId, setEditingComponentId }}> -->
-  <BasicProfile
-    user="UsuarioX1"
-    email="usuario@correo.com"
-    name="Azner"
-    lastname="Aparicio"
-  />
-  <hr />
-  <PersonalInformation
-    name="Azner"
-    lastname="Aparicio"
-    email="usuario@correo.com"
-    faculty="Facultad de Ingeniería de Sistemas Computacionales"
-    career="Licenciatura en Ingeniería de Sistemas de Información"
-    user_type="Estudiante"
-  />
-  <hr />
-  <MateriasInformation />
-  <!-- </EditContext.Provider> -->
+<LandingPage
+	title='EduVision'
+	page={LandingPages.Home}
+>
+	<h1 class='text-5xl font-extrabold dark:text-white'>Mi Perfil</h1>
+	<BasicProfile
+		user='UsuarioX1'
+		email='usuario@correo.com'
+		name='Azner'
+		lastname='Aparicio'
+	/>
+	<hr />
+	<PersonalInformation
+		name='Azner'
+		lastname='Aparicio'
+		email='usuario@correo.com'
+		faculty='Facultad de Ingeniería de Sistemas Computacionales'
+		career='Licenciatura en Ingeniería de Sistemas de Información'
+		user_type='Estudiante'
+	/>
+	<hr />
+	<MateriasInformation />
 </LandingPage>
